cleanup + don't rerun Infer::eliminate
